requirements.txt
langchain
langgraph
fastapi
uvicorn
pytest
python-dotenv
langchain-google-genai
PyPDF2